requirements.txt
krb5
pyyaml
